Make docs build stricter (#1148)
With this commit we treat warnings as errors when building docs with
`make docs`. We also remove the obsolete docs configuration from tox as
documentation should only be built via the make target. Finally, we
upgrade the RTD theme to its latest version to avoid
`RemovedInSphinx30Warning` when building docs.