Make some document properties extensible (#1163)
The `Documents` domain class currently assumes static corpora and thus
populates all properties eagerly via instance fields. There are cases
however when we want to resolve these attributes lazily because the
respective set of documents is generated at a later stage in the
benchmark and we cannot derive these properties upfront. With this
commit we make the relevant properties extensible so they can be
resolved lazily in subclass implementations.