Add IT tests to test installation according to docs (#1193)
As a followup to #1185 this commit adds a few IT tests to validate the
Rally installation steps in the docs.

We also remove hard coded Python versions from the installation section
of docs instead relying on a checked-in variable to ease maintenance
in the future.

Finally we update the PY38/PY39 versions to the latest patch release
(`3.8.8`, `3.9.2`)