Store relative-time in milliseconds (#1221)
Previously `relative-time` has denoted the relative time in microseconds
since the start of a task. This is inconvenient as all other times are
denoted in milliseconds. Therefore, we will migrate this field to be
denoted in milliseconds as well. In the first step we deprecate the
current field and introduce a new field `relative-time-ms` that is used
for the transition period.

Relates #1198