always set default client timeout to 60 (#1308)
Currently, when no client options are defined, Rally uses a hard-coded default of 60 for the timeout value. If client options are defined, but timeout specifically is omitted, then Rally uses the ES Client's default of 10 for the timeout value.

This commit causes the default timeout behavior to be 60 seconds in all cases, to simplify behavior for the user. We use 60 instead of the elasticsearch-py's default of 10 in order to avoid many timeout errors in e.g. bulks and refreshes when there is a lot of load on the benchmark target.