Update rally-* index template settings to default to Elasticsearch defaults  (#1312)
This commit changes the defaults for the number of shards and replicas used in the Elasticsearch datastore first introduced in #1310:
instead of hardcoding them -- when unset -- to `1p/0r`, we rely on ES defaults[1] and only attempt to override these index settings if they've
been explicitly specified in the `ini` file.

Additionally, we also remove the hardcoded `refresh_interval: 5s` setting, to allow us to take advantage of Elasticsearch's default refresh behavior ([2], [3]).

\[1\]: https://www.elastic.co/guide/en/elasticsearch/reference/current/index-modules.html#_static_index_settings
\[2\]: https://www.elastic.co/guide/en/elasticsearch/reference/current/indices-refresh.html#refresh-api-desc
\[3\]: https://www.elastic.co/guide/en/elasticsearch/reference/current/index-modules.html#dynamic-index-settings