Call the correct functions for `scroll-search` and `paginated-search` operations (#1368)
The `Query` runner services three operation types: `search`,
`paginated-search`, and `scroll-search`. Its `__call__` method dispatches to
the appropriate function for one of these operations based on the
`operation-type` passed in via the `params` dict that it accepts as an argument.

The `operation-type` actually never got passed to this method, however. This
resulted in all three operation types ultimately calling the same function
(`request_body_query`), which was not intentional.

To fix this, we now inject the `operation-type` into the dict returned by the
relevant parameter source when the runner is invoked. We also register
`SearchParamSource` for both `scroll-search` and `paginated-search` operations.