Derive correct artifact name for ARM architecture (#1418)
To download an Elasticsearch artifact, Rally needs to resolve the CPU
architecture of the current system to download the correct distribution
tarball. This does not work currently on Macs with ARM M1 chips. These
are reported as `arm64` but the Elasticsearch artifact uses `aarch64` to
refer to this CPU architecture. Consequently, downloading an artifact on
ARM-based Macs fails unless the user specifies `--target-arch=aarch64`.

With this commit we detect this case and ensure that the correct
artifact can be downloaded. We intentionally change the architecture
specifier only in the component that is responsible for deriving the
download URL for artifacts but keep referring to this CPU architecture
as `arm64` in other parts of the code, just as it is reported by the OS.

Co-authored-by: Quentin Pradet <quentin.pradet@gmail.com>