Make online detection more resilient (#1467)
Rally at a very early startup phase determines whether it has internet
access[^1]. Offline is not a terminal event, it just puts itself in the
offline mode and assumes users have been following the offline
install instructions[^2].

This commit makes the check more resilient to transient network errors,
by increasing retries (by default 3) to 8 and timeout to 10s, up to a total of 90s waiting time.
It also logs the error with the default logging settings, to ease troubleshooting.

[^1]: https://github.com/elastic/rally/blob/1b3e656cf95fe2ee57d291d86cfcddcec09c6c3f/esrally/rally.py#L1054
[^2]: https://esrally.readthedocs.io/en/stable/install.html?highlight=offline#offline-install