Install Rally in PEP 517 mode (#1559)
Simply having a pyproject.toml file with a build-system entry as done in https://github.com/elastic/rally/pull/1548 means that `--no-use-pep517` is going to fail. This pull request integrates the relevant part of https://github.com/elastic/rally/pull/1551 to stop doing that.

We need to merge this before the nightly benchmarks run.