Introduce rally-tracks compatibility testing (#1564)
This commit modifies and extends `pytest-rally` so that it can be invoked from
within the Rally repo. This enables Rally developers and CI jobs to test
Rally changes against arbitrary revisions of local track repositories,
using arbitrary versions of Elasticsearch.

The actual contents of the tests live in the track repository. The
plugin will run any tests found in the `it` subdirectory of the provided track
repository by default, including those that are auto-generated by
`pytest-rally`.

By default, tests will be run against the `master` branch of the track
repository checked out in `$RALLY_HOME/benchmarks/tracks/default` (typically
`rally-tracks`), using a build of the `main` branch of Elasticsearch.

To run tests with these defaults, run the following from the root of this
repository:

`pytest it/track_repo_compatibility`

Here is an example invocation that overrides these defaults:

```
pytest it/track_repo_compatibility \
     --track-repository=/path/to/repo \
     --track-revision=some-branch \
     --distribution-version=8.3.2
```