Refactor `pull_revision` to `pull` (#1587)
The method name pull_revision is not consistent with its behaviour, it
only fetches a specific revision from the remote source, and does not
'pull' it into the local checkout.

This commit makes the behaviour of the method consistent with its name,
by removing the pull_revision method, and instead replacing with with
a call to pull.