Skip building plugins moved to modules (#1624)
Since 8.0, we've converted the repository-azure, repository-gcs and 
repository-s3 plugins into Elasticsearch modules, so that they are 
always included. Whilst adding or removing these plugins still succeeds,
it is now a no-op. 

For now we still need to retain the existing post-install hooks 
(see elastic/rally-teams#76), so this is a first step in gracefully 
skipping any build steps.