Skip response parsing for raw requests (#1643)
With this commit we skip response parsing for raw requests. As users
might leverage the `raw-request` operation type for any kind of request,
it is possible that it produces arbitrarily large responses, which might
introduce client-side bottlenecks. As we don't inspect the response
body, we might as well just use the raw response internally and avoid an
accidental bottleneck due to response parsing.