Document Track Processor Interface (#1613)
Closes #1534 

In this PR, I have done the following:

* Moved advanced topics to their own page
* Renamed each advanced topic to hint at what problem they are attempting to solve (e.g. `Template Language` becomes `Templating Track Files with Jinja2`)
* Documented the TrackProcessor interface, including the pitfall of not having #1257 yet.

A note for reviewers, the "meat" of the change here is at the bottom of `advanced.rst`, which by default is collapsed by GH as being too large. The rest is mostly extracted/moved content and some cleaned up typing in doc strings.