Allow to delete indices related to index templates (#1692)
With this commit we make sure that index patterns are allowed when
deleting indices that are related to index templates (both legacy and
composable templates). Prior to this commit, an attempt to delete a
related index would fail with:

```
Wildcard expressions or all indices are not allowed
```

This is because Elasticsearch requires
`action.destructive_requires_name` to be set to `false` in order to
process delete requests with wildcards.

Relates #1152