Propagate request start/end lazily (#1687)
With this commit we only propagate request start and end time to the
parent context if there actually is one. Previously we unconditionally
propagated the request context which could lead to misleading errors
when there was an open request context but no request was issued. This
can happen when a mandatory parameter is missing and the runner was
throwing exception. That exception would get masked by a `KeyError`
because no request start was present yet and thus obfuscate the root
cause.