Refactor version number checks (#1738)
Serverless Elasticsearch doesn't return a `version.number` field from 
the Info API (/) , which is actually something we rely on somewhat 
commonly throughout the codebase.

To fix this, we need a clearly defined way of determining whether or not 
Rally is talking to a Serverless Elasticsearch, this applies for both 
clients used to target the system under test, as well as any ancillary 
clients like those used by a remote metrics store.

In order to do so, we take two approaches:

- Create a `is_serverless | bool` property on existing ES clients objects
- Refactor all version checks to default to selecting the 'minimum' version 
required for a specific conditional, and only use the `is_serverless` property 
if we need to do something specific for serverless