Inject build_flavor in track templates (#1750)
This commit adds `build_flavor` as Jinja2 internal variable to track
templates, index settings, as well as index, component and composable
templates. The variable is equal to `build_flavor` reported by
Elasticsearch cluster, or `default` if `build_flavor` was not
determined.

The `build_flavor` becomes a reserved name, so any track that had used
it as a parameter name needs to be modified.

The commit also adds explicit error message when reserved variable names
are used as track parameters.