Fix serverless public user in trivial tracks (#1785)
This change  eliminates the following blocking points for Rally public
use against serverless clusters:
* The `driver.Driver.prepare_benchmark()` method calls
`client.wait_for_rest_layer()` which includes `_cluster/health` ES API
call which is not available in serverless.
* The `on_benchmark_stop()` in telemetry is incorrectly trying to stop
all telemetry devices even if they were determined to be unsuitable for
serverless.
* The `DeleteIndex` runner includes a call to
`set_destructive_requires_name()` method which relies on
`_cluster/settings` API which is not available in serverless.

With the fixes in it is now possible to run trivial benchmarks with
`delete-index`, `create-index` and `bulk` operarions.