Handle raw response exceptions (#1805)
Some runners set the context manager to return a 'raw' response to avoid 
the unnecessary client side overhead of deserialising the response body. 

For example, the Bulk runner selectively parses the response body to 
check whether the overall HTTP request contains any failed bulk items, 
but discards the rest of the body.

This commit ensures that we properly handle raw request exceptions, and 
those that return empty bodies.